# docs/swagger.yaml
basePath: /api/v1
definitions:
  core.Admin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  core.Plan:
    type: object
  core.TripLocation:
    properties:
      categorie_label:
        example: Shopping
        type: string
      day:
        example: "1"
        type: string
      introduction:
        example: intro to Place1
        type: string
      latitude:
        example: "11.111"
        type: string
      longtitude:
        example: "11.111"
        type: string
      place_id:
        example: place_id_1
        type: string
      place_label:
        example: Place1
        type: string
      thumbnail_url:
        example: https://example.com/image.jpg
        type: string
      time_location:
        example: 08:00AM
        type: string
    type: object
  core.User:
    type: object
  core.Verification:
    properties:
      email:
        type: string
      otp:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API documentation สำหรับ Fiber + Swagger
  title: TiawPao API Documentation
  version: "1.0"
paths:
  /api/v1/admin/login:
    post:
      consumes:
      - application/json
      description: Login an admin into the system
      parameters:
      - description: Admin Data
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/core.Admin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Admin login
      tags:
      - Admin
  /api/v1/admin/register:
    post:
      consumes:
      - application/json
      description: Register a new admin in the system
      parameters:
      - description: Admin Data
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/core.Admin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new admin
      tags:
      - Admin
  /api/v1/plan/addtriplocation/{id}:
    put:
      consumes:
      - application/json
      description: AddTripLocation
      parameters:
      - description: Plan Id
        in: path
        name: id
        required: true
        type: string
      - description: Place Id
        in: body
        name: plan
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: AddTripLocation
      tags:
      - User
  /api/v1/user/createplan:
    post:
      consumes:
      - application/json
      description: Create Plan
      parameters:
      - description: Plan Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/core.Plan'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Plan
      tags:
      - Plan
  /api/v1/user/genotp/{email}:
    get:
      consumes:
      - application/json
      description: Generate a one-time password (OTP) for the user
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate OTP
      tags:
      - User
  /api/v1/user/getuser/{email}:
    get:
      consumes:
      - application/json
      description: Retrieve user details by email. This route is protected by Firebase
        authentication.
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by email
      tags:
      - User
  /api/v1/user/register:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/core.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/core.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - User
  /api/v1/user/update/{email}:
    put:
      consumes:
      - application/json
      description: Update user data in the system
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/core.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update User data
      tags:
      - User
  /api/v1/user/updateuserplan/{email}:
    put:
      consumes:
      - application/json
      description: Update User Plan by Email
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update User Plan
      tags:
      - User
  /api/v1/user/verifyotp:
    post:
      consumes:
      - application/json
      description: Verify the one-time password (OTP) for the user
      parameters:
      - description: Verification Data
        in: body
        name: verification
        required: true
        schema:
          $ref: '#/definitions/core.Verification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify OTP
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"